rename groupId to com.example
