+++ entra-id-quarkus/src/main/java/com/example/ProfilePage.java
@@ -1,4 +1,4 @@
-package no.kantega;
+package com.example;
 
 import io.quarkus.qute.Template;
 import io.quarkus.qute.TemplateInstance;
